moves native scripts to NextJS scripts
