+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - custom "no data" for set of indicators. [OW-179](https://vizzuality.atlassian.net/browse/OW-179)
 
 ### Changed
+- moves native `scripts` to NextJS scripts.
 - updates `eslint@8.9.0` and other linters and plugins.
 - `next@12.1.0`
 - disables robot crawling for staging and preproduction sites.
