--- scripts/crazyegg.jsx
@@ -1,13 +1,0 @@
-import Head from 'next/head';
-
-export default function CrazyEggScript() {
-  return (
-    <Head>
-      <script
-        type="text/javascript"
-        src="//script.crazyegg.com/pages/scripts/0069/4623.js"
-        async="async"
-      />
-    </Head>
-  );
-}
