+++ scripts/hotjar.tsx
@@ -1,13 +1,12 @@
-import Head from 'next/head';
+import Script from 'next/script';
 
 export default function HotjarScript() {
   return (
-    <Head>
-      <script
-        type="text/javascript"
-        // eslint-disable-next-line react/no-danger
-        dangerouslySetInnerHTML={{
-          __html: `
+    <Script
+      id="hotjar"
+      strategy="lazyOnload"
+      dangerouslySetInnerHTML={{
+        __html: `
           (function(h,o,t,j,a,r){
             h.hj=h.hj||function(){(h.hj.q=h.hj.q||[]).push(arguments)};
             h._hjSettings={hjid:1798352,hjsv:6};
@@ -17,8 +16,7 @@ export default function HotjarScript() {
             a.appendChild(r);
           })(window,document,'https://static.hotjar.com/c/hotjar-','.js?sv=');
         `,
-        }}
-      />
-    </Head>
+      }}
+    />
   );
 }
