+++ scripts/userreport.tsx
@@ -1,13 +1,12 @@
-import Head from 'next/head';
+import Script from 'next/script';
 
 export default function UserReportScript() {
   return (
-    <Head>
-      <script
-        type="text/javascript"
-        // eslint-disable-next-line react/no-danger
-        dangerouslySetInnerHTML={{
-          __html: `
+    <Script
+      id="user-report"
+      strategy="lazyOnload"
+      dangerouslySetInnerHTML={{
+        __html: `
             window._urq = window._urq || [];
             _urq.push(['setGACode', '${process.env.NEXT_PUBLIC_GOOGLE_ANALYTICS}']);
             _urq.push(['initSite', '085d5a65-977b-4c3d-af9f-d0a3624e276f']);
@@ -20,8 +19,7 @@ export default function UserReportScript() {
             s.parentNode.insertBefore(ur, s);
             })();
           `,
-        }}
-      />
-    </Head>
+      }}
+    />
   );
 }
